requirements.txt
torch
torchvision
Cython
pillow
scikit-image
xlwt
tqdm